+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+* `[<Class>]` should not be added twice. [#1](https://github.com/nojaf/telplin/issues/1)
+
 ## 0.1.0 - 30-09-22
 
 ### Added